Refactor interfaces of extractor and loader
