+++ wormhole-features/wormhole-pipeline/src/main/java/com/zergclan/wormhole/pipeline/core/filter/precise/convertor/NameConvertor.java
@@ -18,7 +18,7 @@
 package com.zergclan.wormhole.pipeline.core.filter.precise.convertor;
 
 import com.zergclan.wormhole.data.api.DataGroup;
-import com.zergclan.wormhole.data.api.DataNode;
+import com.zergclan.wormhole.data.api.node.DataNode;
 import com.zergclan.wormhole.pipeline.api.Filter;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
