+++ wormhole-kernel/wormhole-metadata/src/main/java/com/zergclan/wormhole/metadata/core/task/TargetMetaData.java
@@ -37,6 +37,8 @@ public final class TargetMetaData implements MetaData {
 
     private final String table;
 
+    private final boolean transaction;
+
     private final Collection<String> uniqueNodes;
 
     private final Collection<String> compareNodes;
