+++ wormhole-plugin/wormhole-plugin-mysql/src/main/java/com/zergclan/wormhole/plugin/mysql/old/writer/MySQLLoader.java
@@ -17,9 +17,6 @@
 
 package com.zergclan.wormhole.plugin.mysql.old.writer;
 
-import com.zergclan.wormhole.data.api.DataGroup;
-import com.zergclan.wormhole.metadata.core.task.TargetMetaData;
-import com.zergclan.wormhole.plugin.api.Loader;
 import lombok.Setter;
 
 import javax.sql.DataSource;
@@ -28,56 +25,11 @@ import javax.sql.DataSource;
  * Loader for MySQL.
  */
 @Setter
-public final class MySQLLoader implements Loader<DataGroup> {
+public final class MySQLLoader {
 
     private final LoadHandler loadHandler;
 
     public MySQLLoader(final DataSource dataSource) {
         this.loadHandler = new LoadHandler(dataSource);
     }
-    
-    @Override
-    public void load(final DataGroup dataGroup, final TargetMetaData target) {
-        // FIXME load implemented
-    }
-    
-//    private void loaderData() {
-//        JdbcLoadContent jdbcLoadContent = (JdbcLoadContent) loadContent;
-//        Collection<Map<String, Object>> listData = jdbcLoadContent.getLoadData();
-//        Map<String, TableMetadata> tables = jdbcLoadContent.getTables();
-//        for (Map.Entry<String, TableMetadata> entry : tables.entrySet()) {
-//            String table = entry.getKey();
-//            TableMetadata tableMetadata = entry.getValue();
-//            System.out.println("load data into " + table + " start...");
-//            List<Map<String, Object>> insertData = new LinkedList<>();
-//            List<Map<String, Object>> updateData = new LinkedList<>();
-//            for (Map<String, Object> mapData : listData) {
-//                int count = executeSelect(tableMetadata, mapData);
-//                if (count > 0) {
-//                    updateData.add(mapData);
-//                } else {
-//                    insertData.add(mapData);
-//                }
-//            }
-//            if (insertData.size() > 0) {
-//                executeBatchInsert(tableMetadata, insertData);
-//            }
-//            if (updateData.size() > 0) {
-//                executeBatchUpdate(tableMetadata, updateData);
-//            }
-//            System.out.println("load data into " + table + " end...");
-//        }
-//    }
-//
-//    private int executeSelect(final TableMetadata tableMetadata, final Map<String, Object> mapData) {
-//        return loadHandler.executeSelect(tableMetadata, mapData);
-//    }
-//
-//    private void executeBatchInsert(final TableMetadata tableMetadata, final List<Map<String, Object>> insertData) {
-//        loadHandler.executeBatchInsert(tableMetadata, insertData);
-//    }
-//
-//    private void executeBatchUpdate(final TableMetadata tableMetadata, final List<Map<String, Object>> updateData) {
-//        loadHandler.executeBatchUpdate(tableMetadata, updateData);
-//    }
 }
